Nutanix VM Inventory using Nutanix API calls
I have made a couple of additions to the code written by Kees Baggerman. His code can be found here -  http://blog.myvirtualvision.com/2019/03/19/vm-reporting-script-for-nutanix-ahv-with-powershell/

I have changed the code so that the VMdetails call to be run even when no GPU is present. This is to retrieve Cluster UUID and CPU details. Also added VM uuid. The CSV filename now uses timestamp.